moved NavCommandLink to its own package to conform with the new TagDecorator
